requirements.txt
annotated-types==0.7.0
anyio==4.5.2
black==24.8.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic-core==2.27.2
numpy==2.3.0
ollama==0.5.1
openai==1.82.1
pandas==2.3.0
google-generativeai    
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
